+++ .eslintrc.js
@@ -1,3 +1,7 @@
+const OFF = 0;
+const WARN = 1;
+
+/** @type {import('eslint').Linter.Config} */
 module.exports = {
   root: true,
   extends: [
@@ -9,7 +13,14 @@ module.exports = {
     {
       files: ["rollup.config.js"],
       rules: {
-        "import/no-extraneous-dependencies": 0,
+        "import/no-extraneous-dependencies": OFF,
+      },
+    },
+    {
+      files: ["templates/**/*.*"],
+      rules: {
+        "prefer-let/prefer-let": OFF,
+        "prefer-const": WARN,
       },
     },
   ],