+++ examples/basic/app/routes/demos/correct.tsx
@@ -1,3 +1,0 @@
-export default function NiceWork() {
-  return <h1>You got it right!</h1>;
-}