+++ examples/blog-tutorial/mocks/index.js
@@ -1,2 +1,0 @@
-require("esbuild-register/dist/node").register();
-require("./start");