+++ examples/pm-app/app/ui/note.tsx
@@ -1,7 +1,0 @@
-import cx from "clsx";
-export function Note({
-  children,
-  className,
-}: React.PropsWithChildren<{ className?: string }>) {
-  return <div className={cx(className, "ui--note")}>{children}</div>;
-}