+++ scripts/release/constants.ts
@@ -1,4 +1,4 @@
-import { cleanupRef, cleanupTagName } from "./utils";
+import { cleanupRef, cleanupTagName, isNightly } from "./utils";
 
 if (!process.env.DEFAULT_BRANCH) {
   throw new Error("DEFAULT_BRANCH is required");
@@ -18,10 +18,11 @@ if (!process.env.VERSION) {
 if (!/^refs\/tags\//.test(process.env.VERSION)) {
   throw new Error("VERSION must start with refs/tags/");
 }
+if (!process.env.PACKAGE_VERSION_TO_FOLLOW) {
+  throw new Error("PACKAGE_VERSION_TO_FOLLOW is required");
+}
 
 export const [OWNER, REPO] = process.env.GITHUB_REPOSITORY.split("/");
-// this one is optional, nightlies only create a single tag,
-// but stable releases create one for each package
 export const PACKAGE_VERSION_TO_FOLLOW = process.env.PACKAGE_VERSION_TO_FOLLOW;
 export const VERSION = cleanupTagName(cleanupRef(process.env.VERSION));
 export const GITHUB_TOKEN = process.env.GITHUB_TOKEN;
@@ -29,3 +30,5 @@ export const GITHUB_REPOSITORY = process.env.GITHUB_REPOSITORY;
 export const DEFAULT_BRANCH = process.env.DEFAULT_BRANCH;
 export const NIGHTLY_BRANCH = process.env.NIGHTLY_BRANCH;
 export const PR_FILES_STARTS_WITH = ["packages/"];
+export const IS_NIGHTLY_RELEASE = isNightly(VERSION);
+export const AWAITING_RELEASE_LABEL = "awaiting release";