+++ packages/remix-deno/CHANGELOG.md
@@ -1,5 +1,72 @@
 # `@remix-run/deno`
 
+## 1.7.6
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.6`
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.6-pre.0`
+
+## 1.7.5
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.5`
+
+## 1.7.4
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.4`
+
+## 1.7.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.3`
+
+## 1.7.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.2`
+
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.1`
+
+## 1.7.0
+
+### Minor Changes
+
+- We've added a new type: `SerializeFrom`. This is used to infer the
+  ([#4013](https://github.com/remix-run/remix/pull/4013)) JSON-serialized return
+  type of loaders and actions.
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.7.0`
+
+## 1.6.8
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.6.8`
+
 ## 1.6.7
 
 ### Patch Changes